+++ packages/core-modules/test/contracts/modules/ElectionModule/Elect.test.js
@@ -5,8 +5,8 @@ const assertRevert = require('@synthetixio/core-js/utils/assertions/assert-rever
 const { getTime, fastForwardTo } = require('@synthetixio/core-js/utils/hardhat/rpc');
 const { daysToSeconds } = require('@synthetixio/core-js/utils/misc/dates');
 const { bootstrap } = require('@synthetixio/deployer/utils/tests');
-const initializer = require('../../helpers/initializer');
-const { ElectionPeriod } = require('../../helpers/election-helper');
+const initializer = require('../../../helpers/initializer');
+const { ElectionPeriod } = require('../../../helpers/election-helper');
 const { findEvent } = require('@synthetixio/core-js/utils/ethers/events');
 
 describe('ElectionModule (vote)', () => {
