+++ packages/core-modules/test/contracts/modules/ElectionModule/Schedule.test.js
@@ -10,8 +10,8 @@ const {
 } = require('@synthetixio/core-js/utils/hardhat/rpc');
 const { daysToSeconds } = require('@synthetixio/core-js/utils/misc/dates');
 const { bootstrap } = require('@synthetixio/deployer/utils/tests');
-const initializer = require('../../helpers/initializer');
-const { ElectionPeriod, assertDatesAreClose } = require('../../helpers/election-helper');
+const initializer = require('../../../helpers/initializer');
+const { ElectionPeriod, assertDatesAreClose } = require('../../../helpers/election-helper');
 const { findEvent } = require('@synthetixio/core-js/utils/ethers/events');
 
 describe('ElectionModule (schedule)', () => {
