+++ packages/core-modules/test/contracts/modules/ElectionModule/Token.test.js
@@ -5,7 +5,7 @@ const assertRevert = require('@synthetixio/core-js/utils/assertions/assert-rever
 const { getTime } = require('@synthetixio/core-js/utils/hardhat/rpc');
 const { daysToSeconds } = require('@synthetixio/core-js/utils/misc/dates');
 const { bootstrap } = require('@synthetixio/deployer/utils/tests');
-const initializer = require('../../helpers/initializer');
+const initializer = require('../../../helpers/initializer');
 const { findEvent } = require('@synthetixio/core-js/utils/ethers/events');
 
 describe('ElectionModule (token)', () => {
